headers: Fix configure --with-default-msvcrt=crtdll
In mingw-w64-headers/configure.ac is missing case for the crtdll. Usage of
crtdll as CRT library is treated with zero value in __MSVCRT_VERSION__ and
setting additional __CRTDLL__ macro. Setting of __CRTDLL__ is needed also
to mimic compatibility with gcc, which can be configure to target
crtdll.dll by default in which case it predefines __CRTDLL__ macro.

Signed-off-by: Martin Storsjö <martin@martin.st>