headers: Fix configure --with-default-msvcrt=msvcr40d
In mingw-w64-headers/configure.ac is missing case for the msvcr40d. It is
missing because all CRT debug libraries in all versions except for VC++4.0,
have just appended letter "d" after the release name.

VC++ 4.0 CRT debug library is exception, the release library is called
msvcrt40 and debug library is called msvcr40d (without T in the middle).
So the current wildcard for msvcrt40* does not catch it. Fix it.

Signed-off-by: Martin Storsjö <martin@martin.st>