headers: Use a more unique name for the __PASTE macro
The previous name, __PASTE, conflicts with the NSS library, which
contains a header that also uses this name. The NSS header first
defines __PASTE to an expansion similar to ours, but then later
undefines __PASTE again, which breaks any further mingw header.

This fixes building NSS after mingw-w64 commit
8f441a62f1b474a29d61571d0da806193ce7ca1c.

Technically, the issue lies with NSS - the namespace of identifiers
starting with two underscores is reserved for the toolchain, where
user code shouldn't be using it.

However, to pragmatically avoid breaking existing code, use a more
unique name for the define in our headers.

Signed-off-by: Martin Storsjö <martin@martin.st>