crt: Remove @<num> stdcall mangling when processing lib-common/*.def.in files for non-I386 builds
This allows to define in lib-common/*.def.in files symbols with I386
stdcall @<num> suffixes and therefore have one common def file for all
platforms.

Removal of @<num> suffixes is done by following sed regex:

    sed -E 's/^([^ ]+)@[0-9]+( |$)/\1\2/'

Note that dollar character in Makefile is duplicated to prevent treating it
as variable in Makefile.

This regex was checked on following test cases:

    Test@4
    Test1@4 ; comment
    Test2@4 == Test22
    Test3@4 @8
    Test@4@8
    Test@5test@8
    @Test6@8

The result is:

    Test
    Test1 ; comment
    Test2 == Test22
    Test3 @8
    Test@4
    Test@5test
    @Test6

That regex handles comments, ordinal numbers and also leading @ chars which
are part of the symbols.

Signed-off-by: Martin Storsjö <martin@martin.st>