crt: Always include load config in mingwex.
This is mandatory on ARM64EC and also required on AArch64 for ARM64X builds.
Instead of extending the existing logic, unify it by always including it in mingwex.

Fill CF Guard fields with zeroes when CF Guard is not enabled in CRT builds. This
leverages existing LLD load config validation warnings to signal potential issues
if the user enables CF Guard while using a CRT that was not built with it.

In practice, this means load config will always be present in output images when using
LLD. Since ld.bfd does not support load config, it remains unaffected.

Signed-off-by: Jacek Caban <jacek@codeweavers.com>