crt: Move definitions of `__mingw_fe_dfl_env` to 'fesetenv.c'
These macros are to be used with `fesetenv()`, so they belong there.

Having these `selectany` variables in 'fenv.h' causes them to be defined with
external linkage in every source file which includes 'fenv.h'. They can also
be exported inappropriately from a DLL, which either is built with
`--export-all-symbols`, or does not contain explicit `dllexport` and is not
built with a DEF file.

Signed-off-by: LIU Hao <lh_mouse@126.com>