crt: Don't delete implicit intermediate .a/.def files after building
Previously, GNU make would delete the implicit intermediate .a/.def
files after building, see [1] for details.

Not deleting them does inflate the size of the build directory a
little bit, but only very marginally (from 201 to 213 MB for a
build with GCC for x86_64).

By keeping them around, they are easier to inspect if debugging
some aspect of them. (Although I haven't really needed to do
this - this isn't the main motivation.)

By not removing them, a build of mingw-w64-crt doesn't end with
a huge wall of text of make doing "rm <all-implicit-files>",
making it easier to observe the output of the other compile
commands instead.

[1] https://www.gnu.org/software/make/manual/html_node/Chained-Rules.html

Signed-off-by: Martin Storsjö <martin@martin.st>