github: Run test-libcxx on aarch64 too
Now when using a version of llvm-mingw that checks out LLVM 21.x,
the libcxx tests pass on mingw/aarch64 (and mingw/armv7) as well,
so we can include this in the test matrix.

This requires setting CMAKE_CXX_COMPILER_TARGET manually; the LLVM
build system infers the wrong architecture when CMake is an
x86_64 binary. (The windows-11-arm images on Github Actions still
bundle an x86_64 CMake binary.)

(The wrong inferred architecture doesn't affect building libcxx,
but the wrong architecture gets passed as --target= parameter
when running the tests.)

Co-authored-by: Jeremy Drake <jeremyd2019@users.sourceforge.net>
Signed-off-by: Martin Storsjö <martin@martin.st>