github: use new public preview windows-11-arm runner
For the "Run extra tests with crt-test" bit (the last bit of the
patch); we could run the built x86 binaries on the arm runners,
but specifically the msvcrt.dll version of crt-test.exe fails
one test when running emulated. (This is because the x86 emulation
doesn't emulate the full 80 bit precision for x87 math.)

For simplicity, just execute the arm binaries there; we still
have full test coverage as the other ones are executed on x86.

Signed-off-by: Martin Storsjö <martin@martin.st>