winpthreads: remove winpthread_internal.h
This header file declares two internal functions:

- __pth_gpointer_locked, which is also declared in thread.h
- pthread_delay_np_ms, which is defined in thread.c

Declaration of `pthread_delay_np_ms` has been moved to thread.h so that
it is declared in header file associtated with source file it is
defined in.

`pthread_delay_np_ms` has been renamed to `_pthread_delay_np_ms`
to make it explicit that this function is internal.

Signed-off-by: Kirill Makurin <maiddaisuki@outlook.com>
Signed-off-by: LIU Hao <lh_mouse@126.com>