winpthreads: remove unused checks from configure.ac
Currently no source files include config.h, which means that
results of all tests from configure.ac are unused.

Most of the checks were testing for presence of features which are
now available with both mingw-w64 and MSVC.

config.h is now included from each source file in `src` subdirectory
so that new configuration checks can be added in future.

Signed-off-by: Kirill Makurin <maiddaisuki@outlook.com>
Signed-off-by: LIU Hao <lh_mouse@126.com>