crt: stat: Don't remove a trailing '\' if it is a DBCS trail byte
In double-byte character sets, the trail byte of a two-byte character
can be a backslash. If such a two-byte character was at the end of
the pathname, the trailing backslash was incorrectly removed.

The code still removes only one trailing directory separator and thus
stat("directory//", &st) still incorrectly fails with MSVCRT. This
commit only fixes the DBCS issue.

Signed-off-by: Martin Storsjö <martin@martin.st>