crt: stat: Fail if the pathname ends with a slash and isn't a directory
POSIX requires stat("foo/", &st) to fail if "foo" isn't a directory or
a symbolic link to a directory. See ENOTDIR in [1].

The statXX functions, that are used with MSVCRT, already have code
to remove a trailing slash. Extend it to reject non-directories with
ENOTDIR if a trailing slash was removed.

These functions aren't used with UCRT because UCRT handles trailing
slashes better than MSVCRT. One minor issue with UCRT is that "foo/"
results in ENOENT instead of ENOTDIR. This commit doesn't fix it.

Preserve errno when calling free().

[1] https://pubs.opengroup.org/onlinepubs/9799919799/functions/fstatat.html

Signed-off-by: Martin Storsjö <martin@martin.st>