Revert "crt: use GetEnvironmentVariableW in getopt"
This reverts commit 8917aca09469d45a4c0c6fe58b36a8eb5b572836.

WinAPI function GetEnvironmentVariableW() should not be used in CRT code as
it reads the variable from the process block, not from the CRT storage.
CRT storage and process block may be out of the sync, for example when
using the CRT's _environ[] access or the 3rd argument envp[] of the main()
function.

CRT getenv() function for CRT builds always returns correct value, so use
it instead of the WinAPI GetEnvironmentVariableW().

Signed-off-by: Martin Storsjö <martin@martin.st>