github: Test building and running the mingw-w64 tests
For llvm-mingw, this starts out directly with a preexisting
build of llvm-mingw (like the ones used for rebuilding the
mingw-w64 crt on top of an existing toolchain release), so this
job can starts right away without waiting for any dependencies.

For GCC, it uses the same GCC as is cross built in the test
pipeline, so those tests start after the second GCC build stage
is completed.

Signed-off-by: Martin Storsjö <martin@martin.st>