crt: scrnsave: Fix library to work on Win95 and WinNT
mingw-w64 scrnsave library contains logic for Win95 but seems that it does
not work on Win95 at all. And it does not work on WinNT too. It can be
verified with WinSDK sample fractal.c screensaver.

The library does not work because GetSystemMetrics() for SM_XVIRTUALSCREEN,
SM_YVIRTUALSCREEN, SM_CXVIRTUALSCREEN and SM_CYVIRTUALSCREEN is not
supported on Windows NT and Windows 95 systems. See WinAPI documentation:
https://web.archive.org/web/20050827010733/http://msdn.microsoft.com/library/en-us/sysinfo/base/getsystemmetrics.asp

Fix this problem and instead of GetSystemMetrics() use GetClipBox() on the
display device context to get rectangular view of screen which works on
every Windows version. Seems that similar logic was used also by WinSDK
scrnsave.lib library when compiling and linking with Visual C++ 6.0.
In screensaver executable compiled by Visual C++ 6.0 is sequence of
GetDC/GetClipBox/ReleaseDC calls.

Tested with WinSDK sample fractal.c screensaver on Win95. Without this
change the screenserver does not show any window and immediately exits.
With this change it correctly visualize fractal. On WinXP there is no
visible change (worked with and also without the change).

Co-authored-by: LIU Hao <lh_mouse@126.com>
Signed-off-by: LIU Hao <lh_mouse@126.com>