crt: tls: Check for errors in tls_atexit's callback for DLL_PROCESS_ATTACH
Now when the tls_atexit's callback is not called for DLL builds as
PIMAGE_TLS_CALLBACK but from DLL entry point, it has an ability to return
failure back to the caller which is LoadLibary or operating system.

TlsAlloc() may fail. Check for failures in tls_atexit's callback, change
callback return value from void to WINBOOL and propagate error back to the
caller.

For EXE builds there is no other option than to continue using the
PIMAGE_TLS_CALLBACK, so on failure call _amsg_exit(16) which shows the
fatal error with information "not enough space for thread data" and exists
process.

Signed-off-by: LIU Hao <lh_mouse@126.com>