github: Stop manually installing Ninja
Since https://github.com/actions/runner-images/pull/11706,
https://github.com/actions/runner-images/pull/11707 and
https://github.com/actions/runner-images/pull/11737, ninja is
preinstalled in the Ubuntu, macOS and x86 Windows runner images.

As noted in https://github.com/actions/partner-runner-images/issues/96,
ninja now also seems to be available on the Windows on ARM
images as well - thus remove redundant installs of ninja on Windows.

This should improve the reliability of the CI pipeline, as the
install of Ninja sometimes fails spuriously, due to Chocolatey
servers not responding.

Signed-off-by: Martin Storsjö <martin@martin.st>