github: Run llvm-mingw tests on aarch64 with msvcrt.dll as well
Initially, we'd skip bundling the runtimes targeting msvcrt
for arm/aarch64, as there's no llvm-mingw prebuilt toolchains
for arm/aarch64 with that configuration.

We can still add the rebuilt target runtimes, targeting msvcrt, on
top of a toolchain that itself is built with UCRT though.

Previously, the aarch64/arm64ec msvcrt configuration was compile
tested by the x86 toolchains, where the llvm-mingw run-tests.sh
scripts test compilation for all architectures, even ones it can't
execute. However, those binaries were never test executed anywhere.

By building aarch64 toolchains that target msvcrt too, we get
testing of executing those binaries as well.

This requires preserving the original DLLs (libc++.dll, libunwind.dll)
that the toolchain itself is linked against, rather than updating
with the newly built runtimes. This, because if building msvcrt
runtimes, with a toolchain that is built for UCRT, we can't replace
the UCRT libc++.dll with a msvcrt version of it; they're not ABI
compatible.

Signed-off-by: Martin Storsjö <martin@martin.st>