crt: correctly handle NULL arguments to mbrtowc
When first argument to mbrtowc is NULL, mbrtowc must behave as
usual except that it produces no output. Before this commit,
it would not update conversion state.

When second argument to mbrtowc is NULL, it must behave if
it has been called as mbrtowc (NULL, "", 1, state). Before this
commit it would simply set `state` to initial conversion state.

Signed-off-by: Kirill Makurin <maiddaisuki@outlook.com>
Signed-off-by: LIU Hao <lh_mouse@126.com>