winpthreads: Reduce the number of threads created in a test on 32 bit
This test repeatedly creates many detached threads, to see if
the pthread_t values end up repeated/reused or not.

While the threads are created detached and the threads exit very
soon after being created, this can still end up creating all
threads in parallel before any of them have exited.

It's practically unfeasible to have 10000 threads in parallel on
32 bit, as it ends up exhausting the virtual address space for the
stacks for the threads.

Therefore, to make the test more reliable in 32 bit mode, lower
the number of tested threads from 10000 to 1000 in 32 bit mode.

This fixes this test when running in 32 bit ARM mode.

Signed-off-by: Martin Storsjö <martin@martin.st>