crt: Disable free space check in mingw-w64 ftruncate64 function
Free space check code in mingw-w64 ftruncate64 function is broken.
For example when the volume is not found, it calls GetDiskFreeSpaceExW()
with NULL first argument which always fails. This code seems to be
unreliable and probably is wrong as the POSIX ftruncate function can
create a file with holes which means that reported file size can be larger
then the underlying disk size. So disable free space check and let followup
WinAPI calls to fail if there are size restriction issues.

Signed-off-by: Martin Storsjö <martin@martin.st>