crt: Fix tmpfile() to work on Windows Vista+ systems
Function tmpfile() in msvcr100 and older versions (including the system os
msvcrt.dll) is trying to create a temporary file in the root of the drive
corresponding to current working directory. In most cases is the current
working directory in some user profile application directory, which means
that the directory for temporary file used by the tmpfile() is often C:\.
Writing to the C:\ is not allowed by unprivileged processes on Windows
Vista and new systems, so tmpfile() is not working at all.

This tmpfile() problem was fixed in msvcr110.dll (and hence also in UCRT)
and tmpfile() in these new versions is creating a temporary file in the
temporary directory returned by WinAPI function GetTempPath() instead of
the root directory.

For affected msvcrt import libraries, replace the msvcrt version of
tmpfile() function by mingw-w64 emulation which attempts to create a
temporary file in GetTempPath() too. For generating unique filename is used
the tmpnam() function together with open(O_CREAT | O_EXCL) in a loop to
prevent race condition in case two different processes or threads generate
same temporary name by the tmpnam() function.

Seems that tmpnam() function generates same file name pattern as the
tmpfile() function. So with this change, the file name should be same, just
located in the proper temporary directory.

Bug: https://sourceforge.net/p/mingw-w64/bugs/921/
Signed-off-by: Martin Storsjö <martin@martin.st>