crt: Document all 32-bit x86 ntdll.dll symbols
Reorder symbols in mingw-w64-crt/lib32/ntdll.def and
mingw-w64-crt/lib-common/ntdllcrt.def.in files and document in
which Windows version was each symbol introduced.

This is just documentation change and does not add or remove any symbol in
32-bit x86 libntdll.a and libntdllcrt.a import libraries.

Newly documented symbols which were missing in the def file were added as
comments.

The fact that this does not change list of symbols can be verified by
comparing lines of ntdll.def file before and after this change that are
same when trimming comments and skipping empty lines. E.g. by command:

    sed 's/\s*;.*//;/^$/d' mingw-w64-crt/lib32/ntdll.def | LC_ALL=C sort

Signed-off-by: LIU Hao <lh_mouse@126.com>