crt: Do not use wide functions in t_fseeko64 test
t_fseeko64.c uses wide char functions for just 7-bit ASCII string.
This is not needed and non-wide functions can be used too.

crtdll.dll does not have _wfopen() and _wtempnam() functions, so this test
cannot be currently compiled for crtdll.dll.

Switch wide functions to non-wide functions and allow compiling this test.

With this change, the t_fseeko64 test is passing also for crtdll.dll and
msvcrt10.dll builds.

Signed-off-by: Martin Storsjö <martin@martin.st>