headers: Silence warning attributes on gets() when building the CRT
Within the implementation of __gets_chk(), we call gets(), in
cases that should be safe (or as safe as they possibly can).

As cases within the build of the CRT itself can be considered
well checked, disable this warning for those cases.

With Clang it is also possible to silence this warning inline with
a #pragma GCC diagnostic ignored "-Wattribute-warning", but GCC
doesn't support any specific -Wno-<diag> option for silencing
an explicit __attribute__((warning)).

Signed-off-by: Martin Storsjö <martin@martin.st>