crt: Don't use -Wpacked
This avoids warnings like this, when building with Clang:

    ../math/fmal.c:34:3: warning: packed attribute is unnecessary for 'struct x87reg_::(anonymous at ../math/fmal.c:34:3)' [-Wpacked]

There seem to be some discrepancies between GCC and Clang, for
which cases they warn about where the packed attribute is
unnecessary.

While the packed attribute can be removed in these cases without
actually affecting the layout of the bitfield structs/unions,
removing the attribute can change the size and alignment of the
struct. Neither of them should make any difference in these
cases though.

But as this warning isn't included in GCC/Clang's -Wextra, just
avoid manually adding it to our list of warning options.

Signed-off-by: Martin Storsjö <martin@martin.st>