crt/stdio: Fix `%.d` output of printf128 for 0
When `%.d` is passed an argument of 0, it should not produce any digits. The
printf128 path uses an intermediate numeric string, which, in this case,
shall be truncated to an empty string. Old code left least one digit so it
erroneously produced "0".

This error was observed in one test in MPFR.

Signed-off-by: LIU Hao <lh_mouse@126.com>