crt: tls: Mark __xl_b, __xl_c and __xl_d variables as static
All those variables are used only in the specified source file and are not
suppose to be used outside of it.

All those variables are already defined via the _CRTALLOC() macro, which
automatically adds __attribute__((used)) keyword, and so these variables
would not be discarded by LTO compiler, even they are unused.

Signed-off-by: LIU Hao <lh_mouse@126.com>