crt: Ensure that stderr is not fully buffered
CRT libraries other than the UCRT can open stderr in full-buffering
mode. This happens for example when output goes to a pipe. The C
standard disallow such mode on stderr (C23 7.23.3 pt. 7) [1]:

> as initially opened, the standard error stream is not fully buffered

Here we ensure that stderr is unbuffered. Note that we can't use line
buffering since it's the same as full buffering [2]:

> _IOLBF: For some systems this mode provides line buffering.
  However on Win32 the behavior is the same as _IOFBF - Full Buffering.

References:

1. https://www.open-std.org/jtc1/sc22/wg14/www/docs/n3220.pdf#subsection.7.23.3
2. https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/setvbuf?view=msvc-170
3. https://sourceforge.net/p/mingw/mailman/message/27121137/

Signed-off-by: Luca Bacci <luca.bacci@outlook.com>
Signed-off-by: LIU Hao <lh_mouse@126.com>