headers/intrin-impl: Allow constant addresses when referencing segmented memory
This solves a leftover in 19fb46e9e8aaf10bf917266e96362fe2b4414934. Because we
cannot pass a segmented memory reference to assembly, the (absolute) address was
always passed in a register, preventing optimization when it could be constant,
for example, in `NtCurrentTeb()`:

   # -masm=att
   movl    $48, %eax
   movq %gs:(%eax), %rax

   # -masm=intel
   mov     eax, 48
   mov rax, gs:[eax]

With this commit, a constant address is allowed to be passed as a constant to
inline assembly. There are two cases:

1. In AT&T syntax, there are different ways to say whether a constant is an
   immediate or (part of) an address. For our use, we must write the operand
   in its address form, using the `a` modifier: `mov %%gs:%a1, %0`

2. In Intel syntax, the `a` modifier on a constant would produce a `ds:`
   segment prefix, so we must not use that. On the other hand, the address form
   of any kind of operand is actually unified, so we enclose the operand in
   brackets, to indicate an address: `mov %0, gs:[%1]`

`NtCurrentTeb()` now produces:

   # -masm=att
   movq %gs:48, %rax

   # -masm=intel
   mov rax, gs:[48]

Signed-off-by: LIU Hao <lh_mouse@126.com>