winpthreads: Implement EXCEPTION_SET_THREAD_NAME for non-i386 builds
For non-i386 builds use asm .seh_handler directive for registering the SEH
handler. This directive register SEH handler for the whole function and
therefore function has to be marked as noinline. Use exactly same SEH
handler as it used for i386 builds, function API is same across platforms.

This asm .seh_handler directive requires support for compiler, assembler
and linker. So check for SEH_INLINE_ASM macro in the same way as it is in
the mingw-w64-crt/crt/crtexe.c file.

Signed-off-by: LIU Hao <lh_mouse@126.com>