winpthreads: Remove usage of global Vectored Exception Handler
Vectored Exception Handler is global for all threads and always active.
Currently it is used and needed only for SetThreadName() function which is
called only from public pthread_setname_np() function and it is not used
very often.

Now when the SetThreadName() function has support for local SEH handler on
all platforms and compilers, there is no need to register global VEH
handler for all threads. So completely remove VEH code.

Signed-off-by: LIU Hao <lh_mouse@126.com>